requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pydantic==2.10.6
pydantic[email]==2.10.6
python-dotenv==1.0.1
google-generativeai==0.8.4
